+++ src/core/inc/mem.h
@@ -41,40 +41,40 @@ enum AS_SEC {
 typedef struct {
     page_table_t pt;
     enum AS_TYPE type;
-    uint64_t colors;
-    uint64_t id;
+    size_t /*uint64_t*/ colors;
+    size_t /*uint64_t*/ id;
     spinlock_t lock;
 } addr_space_t;
 
 typedef struct {
-    uint64_t base;
+    size_t /*uint64_t*/ base;
     size_t size;
-    uint64_t colors;
+    size_t /*uint64_t*/ colors;
 } ppages_t;
 
 struct mem_region {
-    uint64_t base;
+    size_t /*uint64_t*/ base;
     size_t size;
-    uint64_t colors;
+    size_t /*uint64_t*/ colors;
     bool place_phys;
-    uint64_t phys;
+    size_t /*uint64_t*/ phys;
 };
 
 struct dev_region {
-    uint64_t pa;
-    uint64_t va;
+    size_t /*uint64_t*/ pa;
+    size_t /*uint64_t*/ va;
     size_t size;
     size_t interrupt_num;
     uint64_t *interrupts;
-    uint32_t id; /* bus master id for iommu effects */
+    size_t /*uint64_t*/ id; /* bus master id for iommu effects */
 };
 
 typedef struct shmem {
-    uint64_t size;
-    uint64_t colors;
+    size_t /*uint64_t*/ size;
+    size_t /*uint64_t*/ colors;
     bool place_phys;
-    uint64_t phys;
-    uint64_t cpu_masters;
+    size_t /*uint64_t*/ phys;
+    size_t /*uint64_t*/ cpu_masters;
 } shmem_t;
 
 static inline ppages_t mem_ppages_get(uint64_t base, uint64_t size)
@@ -88,19 +88,19 @@ static inline bool all_clrs(uint64_t clrs)
            ((clrs & ((1ULL << COLOR_NUM) - 1)) == ((1ULL << COLOR_NUM) - 1));
 }
 
-void mem_init(uint64_t load_addr, uint64_t config_addr);
-void as_init(addr_space_t* as, enum AS_TYPE type, uint64_t id, void* root_pt,
-             uint64_t colors);
+void mem_init(size_t load_addr, size_t config_addr);
+void as_init(addr_space_t* as, enum AS_TYPE type, size_t id, void* root_pt,
+             size_t colors);
 void* mem_alloc_page(size_t n, enum AS_SEC sec, bool phys_aligned);
-ppages_t mem_alloc_ppages(uint64_t colors, size_t n, bool aligned);
+ppages_t mem_alloc_ppages(size_t colors, size_t n, bool aligned);
 void* mem_alloc_vpage(addr_space_t* as, enum AS_SEC section, void* at,
                       size_t n);
 void mem_free_vpage(addr_space_t* as, void* at, size_t n, bool free_ppages);
 int mem_map(addr_space_t* as, void* va, ppages_t* ppages, size_t n,
-            uint64_t flags);
+            size_t flags);
 int mem_map_reclr(addr_space_t* as, void* va, ppages_t* ppages, size_t n,
-                  uint64_t flags);
-int mem_map_dev(addr_space_t* as, void* va, uint64_t base, size_t n);
+                  size_t flags);
+int mem_map_dev(addr_space_t* as, void* va, size_t base, size_t n);
 
 /* Functions implemented in architecture dependent files */
