+++ src/lib/bit.c
@@ -15,7 +15,7 @@
 
 #include <bit.h>
 
-size_t bit_ctz(uint64_t n)
+size_t bit_ctz(size_t n)
 {
     int i = 0;
 
@@ -26,7 +26,7 @@ size_t bit_ctz(uint64_t n)
     return i;
 }
 
-size_t bit_clz(uint64_t n)
+size_t bit_clz(size_t n)
 {
     int i = 0, j;
