+++ src/lib/inc/bit.h
@@ -28,31 +28,31 @@
 
 #ifndef __ASSEMBLER__
 
-size_t bit_ctz(uint64_t n);
-size_t bit_clz(uint64_t n);
+size_t bit_ctz(size_t n);
+size_t bit_clz(size_t n);
 
-static inline uint64_t bit_get(uint64_t word, uint64_t off)
+static inline size_t bit_get(size_t word, size_t off)
 {
     return word & (1UL << off);
 }
 
-static inline uint64_t bit_set(uint64_t word, uint64_t off)
+static inline size_t bit_set(size_t word, size_t off)
 {
     return word |= (1UL << off);
 }
 
-static inline uint64_t bit_clear(uint64_t word, uint64_t off)
+static inline size_t bit_clear(size_t word, size_t off)
 {
     return word &= ~(1UL << off);
 }
 
-static inline uint64_t bit_extract(uint64_t word, uint64_t off, uint64_t len)
+static inline size_t bit_extract(size_t word, size_t off, size_t len)
 {
     return (word >> off) & BIT_MASK(0, len);
 }
 
-static inline uint64_t bit_insert(uint64_t word, uint64_t val, uint64_t off,
-                                  uint64_t len)
+static inline size_t bit_insert(size_t word, size_t val, size_t off,
+                                  size_t len)
 {
     return (~BIT_MASK(off, len) & word) | ((BIT_MASK(0, len) & val) << off);
 }
