+++ src/lib/printk.c
@@ -22,11 +22,11 @@
 
 #define PRINT_TEXT_LEN 0x100
 
-uint64_t vsprintk(char *buf, const char *fmt, va_list args)
+size_t vsprintk(char *buf, const char *fmt, va_list args)
 {
     char *str;
     str = buf;
-    uint64_t len = strnlen(fmt, PRINT_TEXT_LEN);
+    size_t len = strnlen(fmt, PRINT_TEXT_LEN);
     for (; *fmt; ++fmt) {
         if ((*fmt != '%') && (*fmt != '\n') && (*fmt != '\t')) {
             *str++ = *fmt;
@@ -35,9 +35,9 @@ uint64_t vsprintk(char *buf, const char *fmt, va_list args)
 
         if (*fmt == '%') {
             ++fmt;
-            uint64_t is_unsigned = 0;
-            uint64_t zero_padding = 0;
-            uint64_t is_long = 0;
+            size_t is_unsigned = 0;
+            size_t zero_padding = 0;
+            size_t is_long = 0;
 
             if (*fmt == '0') {
                 ++fmt;
@@ -51,7 +51,7 @@ uint64_t vsprintk(char *buf, const char *fmt, va_list args)
 
             switch (*fmt) {
                 case 'x': {
-                    uint64_t number = is_long ? va_arg(args, uint64_t) : va_arg(args, uint32_t);
+                    size_t number = is_long ? va_arg(args, size_t) : va_arg(args, uint32_t);
                     int length = is_long ? 16 : 8;
                     int length_in_bits = is_long ? 64 : 32;
                     int byte = 0;
@@ -79,20 +79,20 @@ uint64_t vsprintk(char *buf, const char *fmt, va_list args)
                     is_unsigned = 1;
                 case 'i':
                 case 'd': {
-                    uint64_t i, j, max_num_zeros, num_of_digits_uint64_t,
+                    size_t i, j, max_num_zeros, num_of_digits_uint64_t,
                         number, divisor_value_uint64_t,
                         new_div_val = 1, sw_quotient_value = 0;
                     bool keep_zeros = false;
 
                     if (!is_unsigned) {
-                        int64_t number_signed = is_long ? va_arg(args, int64_t) : va_arg(args, int32_t);
+                        long int number_signed = is_long ? va_arg(args, long int) : va_arg(args, int32_t);
                         if (number_signed < 0) {
                             *str++ = 0x2d;
                             number_signed = -(number_signed);
                         }
                         number = number_signed;
                     } else {
-                        number = is_long ? va_arg(args, uint64_t) : va_arg(args, uint32_t);
+                        number = is_long ? va_arg(args, size_t) : va_arg(args, uint32_t);
                     }
 
                     divisor_value_uint64_t = 1000000000;
@@ -164,10 +164,10 @@ uint64_t vsprintk(char *buf, const char *fmt, va_list args)
     return strnlen(fmt, PRINT_TEXT_LEN);
 }
 
-uint64_t printk(const char *fmt, ...)
+size_t printk(const char *fmt, ...)
 {
     va_list args;
-    uint64_t i;
+    size_t i;
 
     char print_buffer[PRINT_TEXT_LEN];
